Defects and wording changes
