+++ HelpMyStreetFE/HelpMyStreetFE/Views/Account/Login.cshtml
@@ -15,7 +15,7 @@
                 <form id="login-form" onsubmit="return false;">
                     <div class="input">
                         <label>Email</label>
-                        <input name="email" type="text" id="email" placeholder="Password" />
+                        <input name="email" type="text" id="email" placeholder="Email" />
                         <span class="error"></span>
                     </div>
                     <div class="input">
