+++ HelpMyStreetFE/HelpMyStreetFE/Views/Account/Verification/_StepCounter.cshtml
@@ -5,7 +5,7 @@
             @Model.Step
         </div>
         <div class="stepper__step__text">
-           Step @Model.Step: @Model.Description
+           @Model.Description
         </div>
     </div>
 </div>
