+++ HelpMyStreetFE/HelpMyStreetFE/Views/Account/Verification/_StepOne.cshtml
@@ -31,7 +31,7 @@
 
 <div class="row">
     <div class="sm12">
-    <partial name="_StepCounter" model='new StepCounterViewModel { Step = 1, Description = "Please have an ID in your name (either Passport or Drivers Licence) and a smartphone (iOS or Android)" }'>
+    <partial name="_StepCounter" model='new StepCounterViewModel { Step = 1, Description = "Please have an ID in your name (either Passport or Driving Licence) and a smartphone (iOS or Android)" }'>
     </div>
 </div>
