+++ HelpMyStreetFE/HelpMyStreetFE/Views/Account/Verification/_StepThree.cshtml
@@ -22,13 +22,18 @@
 
 <div class="row">
     <div class="sm7">
-        <partial name="_StepCounter" model='new StepCounterViewModel { Step = 3, Description = "Scan The Code with the Yoti App" }'>
-            <div class="verification-step__text">
-                <p class="verification-step__text__copy">All our volunteers are ID Verified with our partner Yoti. to provide peace of mind to those in need, as well as their friends and family</p>
-            </div>
+        <div  class="verification-step__yoti__auth__desktop"><partial name="_StepCounter" model='new StepCounterViewModel { Step = 3, Description = "Scan The Code with the Yoti App" }'></div>
+        <div class="verification-step__yoti__auth__mobile"><partial name="_StepCounter" model='new StepCounterViewModel { Step = 3, Description = "Click the blue \"Use Yoti\" Button below" }'></div>
+        <div class="verification-step__text">
+            <p class="verification-step__text__copy verification-step__yoti__auth__desktop">After scanning the code in the app, you will be asked for your consent to share your name and date of birth with HelpMyStreet to verify your identity.</p>
+            <p class="verification-step__text__copy verification-step__yoti__auth__mobile">You will be taken to the Yoti app where you will be asked for your consent to share your name and date of birth with HelpMyStreet to verify your identity. </p>
+        </div>
     </div>
     <div class="sm5">
         <div class="yoti__auth__button" id="@Model.DomId"></div>
+        <div class="text-center verification-step__yoti__auth__desktop">
+            <a id="reload-yoti">Can't see QR code?</a>
+        </div>
     </div>
 </div>
