+++ HelpMyStreetFE/HelpMyStreetFE/Views/Account/Verification/_StepTwo.cshtml
@@ -5,8 +5,8 @@
     <div class="sm8">
         <div class="verification-step__text">
             <partial name="_StepCounter" model='new StepCounterViewModel { Step = 2, Description = "Follow the instructions in the Yoti app to upload your passport or Driving License" }'>
-                <p class="verification-step__text__header--small">You will need to have an ID in your name (either Passport or Driver's Licence)
-                <p>
+                <p class="verification-step__text__header--small">You will need to have an ID in your name (either Passport or Driving Licence)
+<p>
                     <partial name="_YotiInstructions" />
         </div>
     </div>
