+++ HelpMyStreetFE/HelpMyStreetFE/Views/Account/Verification/_VerificationLanding.cshtml
@@ -49,8 +49,8 @@ else
                         <img src="~/img/icons/anonymous-user.png" width="120" />
                     </div>
                     <div class="verification-step__status__title">
-                        You're nearly there <br />
-                        Just your ID Verification to complete!
+                        You're nearly there! <br />
+                        Just your ID Verification to complete.
                     </div>
                 </div>
             </div>
