+++ HelpMyStreetFE/HelpMyStreetFE/Views/Account/Verification/_YotiDownload.cshtml
@@ -1,8 +1,8 @@
 ﻿<div class="verification-step__yoti__download">
     <a href="https://itunes.apple.com/us/app/yoti/id983980808?ls=1&amp;mt=8" target="_blank">
-        <img alt="iOS Store" src="https://www.yoti.com/wp-content/themes/yoti/assets/images/app-store-badge.svg"  width="165" />
+        <img class="verification-step__yoti__download--app" alt="iOS Store" src="https://www.yoti.com/wp-content/themes/yoti/assets/images/app-store-badge.svg"  />
     </a>
     <a href="https://play.google.com/store/apps/details?id=com.yoti.mobile.android.live" target="_blank" >
-        <img alt="Google Play Store" src="https://www.yoti.com/wp-content/themes/yoti/assets/images/google-play-badge.svg" width="180" />
+        <img class="verification-step__yoti__download--play"alt="Google Play Store" src="https://www.yoti.com/wp-content/themes/yoti/assets/images/google-play-badge.svg"  />
     </a>    
 </div>
