+++ HelpMyStreetFE/HelpMyStreetFE/Views/Shared/_AccountNavigation.cshtml
@@ -12,10 +12,10 @@
                 <a href="/account/streets" class="@(Model.CurrentPage == MenuPage.MyStreets ? "selected" : string.Empty)">My Streets</a>
             </li>
         }
-        @if (Model.UserDetails.IsVerified) { 
-            <li class="account__nav__item">
-                <a href="/account/comingSoon" class="@(Model.CurrentPage == MenuPage.ComingSoon ? "selected" : string.Empty)">Coming Soon...</a>
-            </li>
-         }
+   
+        <li class="account__nav__item">
+            <a href="/account/comingSoon" class="@(Model.CurrentPage == MenuPage.ComingSoon ? "selected" : string.Empty)">Coming Soon...</a>
+        </li>
+         
     </ul>
 </nav>
