+++ HelpMyStreetFE/HelpMyStreetFE/Views/Shared/_Layout.cshtml
@@ -89,8 +89,7 @@
 
     <script src="~/lib/jquery/dist/jquery.min.js"></script>
     <script src="~/lib/GDPRCookie/jquery.ihavecookies.min.js"></script>
-    <script src="~/lib/pikaday/moment.js"></script>
-    <script src="~/lib/pikaday/pikaday.js"></script>
+    <script src="~/lib/pikaday/moment.js"></script>    
     <script src="~/main.js" asp-append-version="true"></script>
     @RenderSection("Scripts", required: false)
 </body>
