+++ HelpMyStreetFE/HelpMyStreetFE/src/sass/main.scss
@@ -5,10 +5,8 @@
 @import "./footer.scss";
 @import "./spacing.scss";
 @import "./input.scss";
-@import "./pikaday.scss";
 @import "./registration.scss";
 @import "./login.scss";
-@import "./yoti.scss";
 @import "./forgotton-password.scss";
 @import "./ui/buttons";
