--- HelpMyStreetFE/HelpMyStreetFE/src/sass/yoti.scss
@@ -1,66 +1,0 @@
-
-$screen-md-min: 767px;
-
-@mixin li-yoti {
-    margin-top:18px;
-    font-size: 16px;
-
-}
-
-@mixin ol-yoti {
-    > li {
-        font-weight: bold;
-        counter-increment: item;
-        list-style-type: none;
-        margin: 0;
-        text-indent: -0.7rem;
-        margin-top:15px;
-
-        &::before {
-            content: counter(item) '.';
-            display: inline-block;
-            font-weight: bold;
-            padding-right: 8px;
-        }
-    }
-
-    ul {
-        list-style-type: disc;
-
-        li {
-            padding-left: 8px;
-        }
-    }
-
-
-    li:first-child {
-        margin: 5px 0px 5px 0px
-    }
-
-
-
-    li.no__list__style {
-        &::before {
-            content: none;
-        }
-
-        list-style-type: none;
-        counter-increment: none;
-        margin-top: 5px;
-    }
-}
-
-.verification {
-
- 
-    &__yoti{
-
-        &__confirm {
-            p {
-                margin: 3px 0px 30px 0px;
-            }
-
-        }
-    }
-}
-
